==> Experiments/03_api_knn_vectorized_ngram_2_2.ipynb <==
# Repertoire

try:
  from google.colab import drive
  drive.mount('/gdrive')
except:
  print("vous n'êtes encore  sur Google Colab")

REP_PROJET = '/gdrive/My Drive/Colab Notebooks/Dedoublement/'
REP_INPUT  = REP_PROJET + 'Input/'
REP_INTERMED  = REP_PROJET + 'Intermed/'
REP_OUTPUT  = REP_PROJET + 'Output/'

# MODULES

%cd '/gdrive/My Drive/Colab Notebooks/Dedoublement/'
!pwd

%%capture
pip install -r requirements.txt

import pickle
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics.pairwise import cosine_similarity
from sklearn.neighbors import NearestNeighbors
import numpy as np

# FUNCTIONS

%run "00_functions.ipynb"

# Discogs data

## Load 

baseline = pd.read_pickle(REP_INTERMED+ "baseline.sav")

## Vectorizer :  Keep TF-IDF strategy 

start_time = datetime.now()

create_vectorizer_and_vectorized_db(baseline, analyzer = 'char',ngram_range=(2,2))

end_time = datetime.now()

print('Duration: {}'.format(end_time - start_time))

# Duration: 02 min:15s

analyzer = 'char'
ngram_range = (2,2)
X_train, vectorizer = reload_data_and_vectorizer(analyzer = analyzer, ngram_range=ngram_range)

# Query Data

df_no_match = pd.read_pickle(REP_INTERMED +'base_cdandlp.sav').reset_index(drop=True)

df_no_match['text'] = df_no_match['artiste'] +  ' ' + df_no_match['titre']

df_sample = df_no_match.sample(100, random_state=20).reset_index(drop=True)

del df_no_match

# KNN unsupervised (sklearn api)

## Training with Manahattan


suffix_gram = re.sub(r"[(,)]", "", str(ngram_range)).replace(' ','_')

path = REP_INTERMED + analyzer + '_' + suffix_gram+  '_results_sim_Manhattan.pkl'

nbrs = NearestNeighbors(n_neighbors=100, metric="manhattan",algorithm='brute').fit(X_train)

start_time = datetime.now()

result = [ nbrs.kneighbors(text_to_vect(df_sample.iloc[i,3], vectorizer)) for i in range (len(df_sample)) ]

f = open(path, 'wb')

pickle.dump(result, f) 

f.close()

end_time = datetime.now()

print('Duration: {}'.format(end_time - start_time))

# Customiz Output

suffix_gram = re.sub(r"[(,)]", "", str(ngram_range)).replace(' ','_')
analyzer = 'char'
ngram_range = (2,2)

path = REP_INTERMED + analyzer + '_' + suffix_gram+  '_results_sim_Manhattan.pkl'

with open(path, 'rb') as file:
      result = pickle.load(file)

temp = [ customize_output(i,results= result, df_baseline=baseline.copy(), sample=df_sample) for i in range(len(df_sample)) ]

pd.concat(temp,axis=0).\
    rename(columns={'master_id':'id_text'}).\
    to_excel(REP_INTERMED + analyzer + '_' + suffix_gram +'_neighbors_Manhatthan.xlsx', index= False)

pd.concat(temp,axis=0)
