==> Experiments/04_Blocking_country.ipynb <==
# Repertoire

try:
  from google.colab import drive
  drive.mount('/gdrive')
except:
  print("vous n'êtes encore  sur Google Colab :(")

REP_PROJET = '/gdrive/My Drive/Colab Notebooks/Dedoublement/'
REP_INPUT  = REP_PROJET + 'Input/'
REP_INTERMED  = REP_PROJET + 'Intermed/'
REP_OUTPUT  = REP_PROJET + 'Output/'

# MODULES

%cd '/gdrive/My Drive/Colab Notebooks/Dedoublement/'
!pwd

%%capture
! pip install -r requirements.txt

import numpy as np
import pandas as pd
import pickle
from datetime import datetime
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics.pairwise import cosine_similarity
from sklearn.neighbors import NearestNeighbors
from functools import partial

# FUNCTIONS

%run "00_functions.ipynb"

# Inputs Discogs

Merge the extraction and aggregate data to artist + title level

baseline = pd.read_pickle(REP_INTERMED + 'baseline.sav').drop([	'released_date'], axis=1)
baseline['country'] = baseline['country'].str.title()

print(" frequence de country manquant en %: " ,round(baseline[baseline['country'] == 'Nan'].shape[0]/baseline.shape[0]),2)

baseline['country'].value_counts().reset_index()

# Recuperer l'ensemble des pays

list_temp = baseline['country'].unique().tolist()
list_temp = [ country.replace('Usa','Us') for country in list_temp]

elements = [ (country.replace('¤' , ',').replace('&' , ',').replace(';' , ',').replace(' ,', ',').replace(', ', ',')).split(',')for country in list_temp ]

list_country  = []
for element in  elements :
  for mot in element :
    list_country.append(mot)
list_country = set(list_country)


list_country.remove('')
#list_country.remove('Nan')

#with open(REP_INTERMED + 'list_country.pkl','wb') as f :
#     pickle.dump(list_country, f)

**!!!!!!!!!!! la cellule sert à batcher les index du dataframe cdandlp par pays! Pas besoin de l'exécution une deuxième fois car l'output sera sauvégarder en pkl pour faciliter sa lecture** Temps d'exécution 2 min

# list_baseline_country =[ baseline[baseline['country'].isin([i, 'Nan'])]  for i in list_country ]
# with open(REP_INTERMED + 'list_baseline_country.pkl','wb') as f :
#      pickle.dump(list_baseline_country, f)

# Inputs CDandLP

Scope : 
+ on articles for which we do not know the correspondence on the discogs (id_release=0) 
+ on articles where a sql query can be used to perfectly match discogs and cdandlp data

## Données brutes

import pandas as pd
#to_drop = ['n_ident',	'n_categ1',	'n_categ2',	'code_upload_not_null',	'courtdesc_lg1',	'courtdesc_lg2',	'format'	,'annee',	'label']
df = pd.read_pickle(REP_INTERMED + 'base_cdandlp.sav')
df['text']  = df['artiste']  + ' ' +  df['titre']
df.head()

print(" frequence de pressage manquant en : " ,round(df[df['pressage'].isna()].shape[0]/df.shape[0],2))

## Data cleaning

df['pressage'] = df['pressage'].str.title()
df['pressage'] = df['pressage'].str.strip()
df.loc[df['pressage'].isna(), 'pressage'] = 'Vide'
df = recode_pressage(df)

## Dict of Dataframes

dict_country_baseline = {}
for country, artist, title, text, master_id in zip(baseline['country'], baseline['artist'], baseline['title'], baseline['text_CLEAN'], baseline['master_id']):
    if country in list_country:
        if country not in dict_country_baseline:
            dict_country_baseline[country] = [ (country, artist, title, text, master_id) ]
        else:
            liste_temp = dict_country_baseline[country]
            liste_temp.append((country, artist, title, text, master_id))
            dict_country_baseline[country] = liste_temp

with open(REP_INTERMED + 'dict_country_baseline.pkl', 'wb') as f:
  pickle.dump(dict_country_baseline, f)  

list_country.remove('Nan')
dict_country_cdandlp = {}
for pressage, text, n_ref, artiste, titre in zip(df['pays_clean'], df['text'], df['n_ref'], df['artiste'], df['titre']):
    if pressage in list_country:
        if pressage not in dict_country_cdandlp:
            dict_country_cdandlp[pressage] = [ (pressage,text,n_ref, artiste, titre) ]
        else:
            liste_temp = dict_country_cdandlp[pressage]
            liste_temp.append((pressage,text, n_ref, artiste, titre))
            dict_country_cdandlp[pressage] = liste_temp  

with open(REP_INTERMED + 'dict_country_cdandlp.pkl', 'wb') as f:
  pickle.dump(dict_country_cdandlp, f)     

**!!!!!!! The cell below write outputs in write**

df_items = pd.concat(
                 [ pd.DataFrame(dict_country_baseline['Nan'], columns=col_baseline), 
                   pd.DataFrame(dict_country_baseline['France'],columns=col_baseline)
                 ], ignore_index=True)

df_sample= pd.DataFrame(dict_country_cdandlp['France'], columns=col_cdandlp)

answer = knn_country(df_items, df_sample)

# Exécutions : prochaine passage

Les codes ci-après permettent de recharger les resultats des étapes précedentes sans avoir à compiler les cellules précedentes.

# tables par pays
col_baseline  = ['country', 'artist', 'title', 'text_CLEAN', 'master_id']
col_cdandlp = ['pays_clean', 'text', 'n_ref', 'artiste', 'titre']

with open(REP_INTERMED + 'dict_country_cdandlp.pkl', 'rb') as f:
  dict_country_cdandlp = pickle.load(f)  


with open(REP_INTERMED + 'dict_country_baseline.pkl', 'rb') as f:
  dict_country_baseline = pickle.load(f)  


df_items = pd.concat(
                      [ pd.DataFrame(dict_country_baseline['France'],columns=col_baseline),
                        pd.DataFrame(dict_country_baseline['Nan'], columns=col_baseline)]
                     , ignore_index=True)

df_sample= pd.DataFrame(dict_country_cdandlp['France'], columns=col_cdandlp)

# fichier resultats
with open(REP_INTERMED + df_items.loc[0,'country'] + '_results_sim.pkl', 'rb') as f:
        answer = pickle.load(f) 

:

tmp = pd.concat(answer)
tmp.head()
