==> Experiments/04_Blocking_words.ipynb <==
# Repertoire

try:
  from google.colab import drive
  drive.mount('/gdrive')
except:
  print("vous n'êtes encore  sur Google Colab :(")

REP_PROJET = '/gdrive/My Drive/Colab Notebooks/Dedoublement/'
REP_INPUT  = REP_PROJET + 'Input/'
REP_INTERMED  = REP_PROJET + 'Intermed/'
REP_OUTPUT  = REP_PROJET + 'Output/'

import pickle
with open(REP_INTERMED + 'dict_words.pkl', 'rb') as f:
        dict_words = pickle.load(f) 

dict_words

# MODULES

%cd '/gdrive/My Drive/Colab Notebooks/Dedoublement/'
!pwd

%%capture
! pip install -r requirements.txt

import numpy as np
import pandas as pd
import pickle
from datetime import datetime
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics.pairwise import cosine_similarity
from sklearn.neighbors import NearestNeighbors
from functools import partial

# FUNCTIONS

%run "00_functions.ipynb"

# Inputs

baseline = pd.read_pickle(REP_INTERMED + 'baseline.sav').drop([	'released_date'], axis=1) 
baseline['text_2'] = baseline['text_CLEAN'].str.title()

df = pd.read_pickle(REP_INTERMED + 'base_cdandlp.sav')
df['text']  = df['artiste']  + ' ' +  df['titre']

# Dict of words

baseline['text_2'] = baseline['text_CLEAN'].str.title()
# stop_words = ['and', 'et']
# dict_words = {}


for id_row,text in enumerate(baseline['text_2']):
    words = text.split()
    words = [ w for w in words if w not in stop_words ]
    words = set(words)
    for word in words:
        if word not in dict_words:
            dict_words[word] = [ id_row ]
        else:
            liste_temp = dict_words[word]
            liste_temp.append(id_row)
            dict_words[word] = liste_temp
  
 with open(REP_INTERMED + 'dict_words.pkl', 'wb') as f:
        pickle.dump(dict_words, f) 

with open(REP_INTERMED + 'dict_words.pkl', 'rb') as f:
        dict_words = pickle.load(f) 

# Modeling

## Based on knn from scratch

start_time = datetime.now()
row = df.loc[0, 'text']
# retrieve the list of indices by word
indices_list = [dict_words.get(key) for key in row.split()]
while None in indices_list:
    indices_list.remove(None)
# convert the list of lists to flat list
indices_list_flat = list(set([item for sublist in indices_list for item in sublist]))
# match indexes with baseline indexes
df_baseline = baseline.iloc[indices_list_flat,].reset_index(drop=True)
answer = knn_words(df_baseline, df.iloc[[0,]])
#pd.concat(answer)
end_time = datetime.now()
print('Duration: {}'.format(end_time - start_time))


answer = knn_words(df_baseline, df.iloc[[0,]])
#pd.concat(answer)
end_time = datetime.now()
print('Duration: {}'.format(end_time - start_time))

start_time = datetime.now()
row = df.loc[2, 'text']
# retrieve the list of indices by word
indices_list = [dict_words.get(key) for key in row.split()]
while None in indices_list:
    indices_list.remove(None)
# convert the list of lists to flat list
indices_list_flat = list(set([item for sublist in indices_list for item in sublist]))
# match indexes with baseline indexes
df_baseline = baseline.iloc[indices_list_flat,].reset_index(drop=True)
df_sample = df.iloc[[2,]].reset_index(drop='True')
answer = knn_words(df_baseline, df_sample)
end_time = datetime.now()
print('Duration: {}'.format(end_time - start_time))
pd.concat(answer)


## Based on text distances

start_time = datetime.now()
row = df.loc[0, 'text']
# retrieve the list of indices by word
indices_list = [dict_words.get(key) for key in row.split()]
indices_list.remove(None)
# convert the list of lists to flat list
indices_list_flat = list(set([item for sublist in indices_list for item in sublist]))
# match indexes with baseline indexes
df_baseline = baseline.iloc[indices_list_flat,].reset_index(drop=True)
df_baseline['text_query'] = make_text_prep_func(row)
df_baseline['index'] = df_baseline.index
# KNN based on text distances
answer2 = knn_words_distance(0, df_baseline, df.iloc[[0,]], 20)
end_time = datetime.now()
print('Duration: {}'.format(end_time - start_time))
answer2

## Based on sklearn knn

start_time = datetime.now()
k=20
row = df_sample.loc[37, 'text']
df_sample=df_sample.iloc[[37,]]
# retrieve the list of indices by word
indices_list = [dict_words.get(key) for key in row.split()]
indices_list.remove(None)
# convert the list of lists to flat list
indices_list_flat = list(set([item for sublist in indices_list for item in sublist]))
# match indexes with baseline indexes
df_baseline = baseline.iloc[indices_list_flat,].reset_index(drop=True)
if len(df_baseline) < k :
  k =len(df_baseline)
# train
vectorizer = TfidfVectorizer(analyzer = 'char', ngram_range=(2,2) , use_idf = False)
X_train = vectorizer.fit_transform(df_baseline['text_CLEAN'].str.replace(' ', ''))
# KNN with sklearn api
nbrs = NearestNeighbors(n_neighbors=k, metric="cosine",algorithm='brute').fit(X_train)
result = nbrs.kneighbors(text_to_vect(df_sample.loc[37,'text'], vectorizer))
# customize output
distances, indices = result
end_time = datetime.now()
print('Duration: {}'.format(end_time - start_time))
customize_output_blocking(37,results= result, df_baseline=df_baseline.copy(), sample=df_sample)

# Level 1 running

col_cdandlp = ['pays_clean', 'text', 'n_ref', 'artiste', 'titre']
with open(REP_INTERMED + 'dict_country_cdandlp.pkl', 'rb') as f:
  dict_country_cdandlp = pickle.load(f)  

df_sample= pd.DataFrame(dict_country_cdandlp['France'], columns=col_cdandlp)

level1_knn_words(37, df_sample, baseline, dict_words)

result = [ level1_knn_words(indice, df_sample, baseline, dict_words)  for indice in range(len(df_sample)) ]

df = df_sample
baseline
index =37
row = df.loc[index, 'text']
  # retrieve the list of indices by word
indices_list = [dict_words.get(key) for key in row.split()]
distances = ['levenshtein', 'jaro_winkler', 'jaccard', 'overlap', 'hamming', 'fuzzy_partial']
while None in indices_list:
      indices_list.remove(None)
# convert the list of lists to flat list
indices_list_flat = list(set([item for sublist in indices_list for item in sublist]))
df_baseline = baseline.iloc[indices_list_flat,].reset_index(drop=True)
df_sample = df.iloc[[index,]].reset_index(drop='True')

def top_k(x, k):
        ind = np.argpartition(x, -1 * k)[-1 * k:]                  
        return ind[np.argsort(x[ind])]                    

topsim = np.apply_along_axis(lambda x: top_k(x, nb_top), 0, mat_sim)
topsim2 = pd.DataFrame( topsim, columns=['index'])

l_sim = []
for i in list(topsim2['index']):
    l_sim.append( float(mat_sim[i])*100.0 )
topsim2['similarity'] = l_sim

nb_top

 
   #recupération des infos d'origine
df_nearest = pd.merge(baseline, topsim2, on = 'index', how='inner')

df_nearest["rank"] =  df_nearest['similarity'].rank(method ='min',ascending = False).astype(int)

df_nearest['id_query'] =sample.loc[indice,'n_ref']

df_nearest['text_query'] = make_text_prep_func(sample.loc[indice,'text'])

distances = ['levenshtein', 'jaro_winkler', 'jaccard', 'overlap', 'hamming', 'fuzzy_partial']

col_list_values = df_nearest.apply(lambda p:compound_similarity(p, "text_CLEAN", "text_query"), axis=1)
    
df_nearest.drop(['index','text_query', 'text_CLEAN'], axis=1, inplace=True)

df_nearest = pd.concat([
                            df_nearest[['id_query', 'master_id', 'artist', 'title', 'similarity', 'rank']], 
                            pd.DataFrame(np.column_stack(col_list_values).T, columns=distances )], 
                          axis= 1
                          )

df_nearest.sort_values(by='similarity', ascending=False, inplace=True)
new_row = pd.DataFrame(np.array([[
                                    sample.loc[indice,  'n_ref'],
                                    sample.loc[indice,  'n_ref'], 
                                    sample.loc[indice, 'artiste'],
                                    sample.loc[indice, 'titre'],
                                    float(-1),
                                    0,
                                    1,
                                    1,
                                    1,
                                    1,
                                    1,
                                    1,
                                  ]]
                                  ), 
                        columns= ['id_query', 'master_id', 'artist', 'title', 'similarity', 'rank'] + distances
                        )

df_nearest = new_row.append(df_nearest, ignore_index=True)

# Fin

"""# Inputs"""
baseline = pd.read_pickle(REP_INTERMED + 'baseline.sav').drop([	'released_date'], axis=1)

with open(REP_INTERMED + 'dict_words.pkl', 'rb') as f:
  dict_words = pickle.load(f)

with open(REP_INTERMED + 'dict_country_cdandlp.pkl', 'rb') as f:
  		dict_country_cdandlp = pickle.load(f)
    
dict_country_cdandlp_france = dict_country_cdandlp['France']


def knn(start , end, baseline = baseline, dict_words=dict_words, dict_country_cdandlp =dict_country_cdandlp_france ) :

  df_sample= pd.DataFrame(dict_country_cdandlp_france[start: end], columns=['pays_clean', 'text', 'n_ref', 'artiste', 'titre'])

  result = [ level1_knn_words(indice, df_sample, baseline, dict_words)  for indice in range(len(df_sample)) ]
  
  with open(REP_INTERMED + str(end) + '_France_words_knn.pkl', 'wb') as  f1:
    pickle.dump(result, f1) 

taille = len(dict_country_cdandlp_france)
pas = int(taille/15)
list_tmp = [ indices for indices in range(0,  taille, pas) ]

 # start
start_time = datetime.now()

[ knn(start, end) for start , end in zip(list_tmp,list_tmp[1:])]

end_time = datetime.now()
print('Duration: {}'.format(end_time - start_time))

  with open(REP_INTERMED + str(1) + '_France_words_knn.pkl', 'rb') as  f1:
    tmp = pickle.load(f1) 

pd.concat(tmp)
